Final commit

+++ 01.12/pipe2.c
@@ -3,35 +3,51 @@
 #include <unistd.h>
 #include <fcntl.h>
 #include <signal.h>
+#include <time.h>
+
 
 int ball = 0;
+time_t *SigIntTime;
+int FirstPress = 1;
 
-int flag = 0;
 int pid;
 void w_sigusr1(int sig){
 	signal(SIGUSR1, w_sigusr1);
 }
 
-void w_sigint2(int sig){
-	kill(pid, SIGKILL);
-}
 
 void w_sigint(int sig){
-	printf("%d\n", ball);
-	signal(SIGINT, w_sigint2);	
+	time_t *CurrentTime = (time_t *) malloc(sizeof(time_t)); 
+	time(CurrentTime);
+		if(FirstPress){
+			SigIntTime = CurrentTime;
+			FirstPress--;
+			if (getpid() == pid)
+				printf("%d\n", ball);
+			signal(SIGINT, w_sigint);
+			return;
+		}
+	if(getpid() == pid)
+		printf("%d\n", ball);
+	if (difftime(*CurrentTime, *SigIntTime) < 1)
+		kill(pid, SIGKILL);
+	SigIntTime = CurrentTime;
+	signal(SIGINT, w_sigint);	
 }
 
 
 int main(int argc, char const *argv[])
 {
+
+	//SigIntTime = (time_t*) malloc(sizeof(time_t));	
 	signal(SIGUSR1, w_sigusr1);
 	signal(SIGINT, w_sigint);
 
 	pid = getpid();
 	int fd[2];
 	pipe(fd);
-	int p = fork();
 	write(fd[1], &ball, sizeof(ball));
+	int p = fork();
 	if (!p){
 		while(1){
 			read(fd[0], &ball, sizeof(ball));
